Added XIO flash support
Added flash read (and write) support to u-boot. For most parts
the pnx8550_nand driver for linux 2.6.21.3 was used. Read is
working as expected, but ECC failes miserably. Seems, that the
ECC's differ eigther between linux <--> WinCE or u-boot <-->
linux. Need more testing, a disassembly of the WinCE algorithm
or best, the c source for it. Also bad block support is -not-
tested!

The u-boot mtd does -not- support neigther partitions nor
read only regions! Because of the current state, writes to
the flash are -not- recommended and will lead to breakage.

Signed-off-by: Laszlo Hegedüs <laszlo.hegedues@gmail.com>