Add support for PR4450
Prelimary support for the pr4450 processor. It is a (almost) mips32
compiliant thing. It's counter0 is not working correctly, so I had
to revert the default time.c to the old one used in 1.2.0.

I will extract the support for the processor from the mips32 (same
as xburst) when I have the time. Will be a bunch of copy & paste
though, since most of the internals are the same.

Signed-off-by: Laszlo Hegedüs <laszlo.hegedues@gmail.com>