New mkimage OS type to accomodate the philips style linux kernel
Adds new OS type to both mkimage and the U-Boot bootloader to
accomodate for the boot parameter passing in the philips bootloader.
This target is used for the backup kernel residing in the flash of
the pollinux board.

Currently this patch is more like a hack. Further improvements are
desired.

Signed-off-by: Laszlo Hegedüs <laszlo.hegedues@gmail.com>